FDS Source: Remove uninitialized variables etc.

+++ Source/main.f90
@@ -463,7 +463,7 @@ MAIN_LOOP: DO
 
    ! Determine when to dump out diagnostics to the .out file
 
-   LO10 = LOG10(REAL(MAX(1,ABS(ICYC)),EB))
+   LO10 = INT(LOG10(REAL(MAX(1,ABS(ICYC)),EB)))
    IF (MOD(ICYC,10**LO10)==0 .OR. MOD(ICYC,100)==0 .OR. (T+DT)>=T_END) DIAGNOSTICS = .TRUE.
 
    ! If evacuation, set up special time iteration parameters
